[Feature][scaleph-ui]flink deploy config file web (#100)
* fix: datasource plugin construct jdbcurl error.

* fix: seatunnel and flink docker image build args.

* feature: change seatunnel base image to ghcr.io/flowerfine/scaleph_flink:1.13.6_2.11

* feature: flink cluster instance dao entity, mapper, xml

* feature: flink cluster instance dto and convert.

* feature: flink cluster instance service.

* feature: flink cluster instance web

* feature: flink cluster instance manage service.

* feature: flink cluster instance manage service.

* feature: flink release upload web.

* feature: flink release batch delete web.

* feature: flink release download web.

* feature: remove unused imports.

* feature: adjust name.

* feature: adjust name.

* feature: flink deploy-config web.

* feature: flink deploy-config web.

* feature: flink deploy-config upload web.

* feature: flink deploy-config new web.

* feature: add flink deploy-config param

* feature: flink deploy-config batch delete.

* feature: flink deploy-config batch delete.

* feature: flink deploy-config list icon.

* feature: flink deploy-config update web.

* feature: flink deploy-config update web.

* feature: flink deploy-config delete web.

+++ scaleph-engine/scaleph-engine-flink/src/main/java/cn/sliew/scaleph/engine/flink/service/impl/FlinkDeployConfigFileServiceImpl.java
@@ -26,7 +26,6 @@ import cn.sliew.scaleph.engine.flink.service.convert.FileStatusVOConvert;
 import cn.sliew.scaleph.engine.flink.service.convert.FlinkDeployConfigFileConvert;
 import cn.sliew.scaleph.engine.flink.service.dto.FlinkDeployConfigFileDTO;
 import cn.sliew.scaleph.engine.flink.service.param.FlinkDeployConfigFileListParam;
-import cn.sliew.scaleph.engine.flink.service.param.FlinkDeployConfigFileUpdateParam;
 import cn.sliew.scaleph.engine.flink.service.vo.FileStatusVO;
 import cn.sliew.scaleph.storage.service.FileSystemService;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
@@ -44,7 +43,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Serializable;
 import java.util.List;
-import java.util.Map;
 
 import static cn.sliew.scaleph.common.exception.Rethrower.checkArgument;
 
@@ -88,10 +86,8 @@ public class FlinkDeployConfigFileServiceImpl implements FlinkDeployConfigFileSe
     }
 
     @Override
-    public int update(FlinkDeployConfigFileUpdateParam param) {
-        final FlinkDeployConfigFile record = new FlinkDeployConfigFile();
-        record.setId(param.getId());
-        record.setRemark(param.getRemark());
+    public int update(FlinkDeployConfigFileDTO dto) {
+        final FlinkDeployConfigFile record = FlinkDeployConfigFileConvert.INSTANCE.toDo(dto);
         return flinkDeployConfigFileMapper.updateById(record);
     }
 
@@ -99,9 +95,12 @@ public class FlinkDeployConfigFileServiceImpl implements FlinkDeployConfigFileSe
     public int deleteById(Serializable id) {
         try {
             final FlinkDeployConfigFileDTO flinkDeployConfigFileDTO = selectOne(id);
-            final List<FileStatus> fileStatuses = fileSystemService.listStatus(getFlinkDeployConfigFileRootPath() + "/" + flinkDeployConfigFileDTO.getName());
-            for (FileStatus fileStatus : fileStatuses) {
-                deleteDeployConfigFile((Long) id, fileStatus.getPath().getName());
+            final String rootPath = getFlinkDeployConfigFileRootPath() + "/" + flinkDeployConfigFileDTO.getName();
+            if (fileSystemService.exists(rootPath)) {
+                final List<FileStatus> fileStatuses = fileSystemService.listStatus(rootPath);
+                for (FileStatus fileStatus : fileStatuses) {
+                    deleteDeployConfigFile((Long) id, fileStatus.getPath().getName());
+                }
             }
             return flinkDeployConfigFileMapper.deleteById(id);
         } catch (IOException e) {
@@ -111,11 +110,11 @@ public class FlinkDeployConfigFileServiceImpl implements FlinkDeployConfigFileSe
     }
 
     @Override
-    public int deleteBatch(Map<Integer, ? extends Serializable> map) {
-        for (Serializable id : map.values()) {
+    public int deleteBatch(List<Long> ids) {
+        for (Long id : ids) {
             deleteById(id);
         }
-        return flinkDeployConfigFileMapper.deleteBatchIds(map.values());
+        return ids.size();
     }
 
     @Override
